Move refactored common unit test code
